DFS fixes minimum spanning
I implemented a new function using DFS instead of BFS to fix the issue BFS was giving where the progam sort of minimum spanning tree. If the agent was at node A the agent would often revisit a large number of node unnecessarily instead of going to the unvisited node nearest to A. DFS gives a massive performance benefit.